pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name="ansys-scade-git"
version="0.1.dev0"
description ="SCADE GUI extensions and command line utilities for Git."
readme="README.rst"

# only 3.7. and 3.10
requires-python = ">=3.7,!=3.8.*,!=3.9.*,<3.11"
license = {file = "LICENSE"}
authors = [
    {name = "Ansys, Inc.", email = "pyansys.core@ansys.com"},
]
maintainers = [
    {name = "Ansys, Inc.", email = "pyansys.core@ansys.com"},
]
dependencies=[
    "ansys-scade-apitools",
    "dulwich>=0.21.3",
    "importlib-metadata >= 1.0; python_version < '3.8'",
    "importlib-metadata >= 4.0; python_version >= '3.8'",
    "lxml",
]
classifiers=[
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows",
]
keywords=["Scade", "Git", "CM", "Merge"]

[project.optional-dependencies]
build = [
    "build==1.2.2.post1",
    "twine==6.0.1"
]
tests = [
    "pytest==8.3.4",
    "pytest-cov==6.0.0"
]
doc = [
    "ansys-sphinx-theme==1.2.4",
    "jupyter_sphinx==0.5.3",
    "nbsphinx==0.9.6",
    "numpydoc==1.8.0",
    "sphinx==8.1.3",
    "sphinx-autodoc-typehints==2.5.0",
    "sphinx-copybutton==0.5.2",
    "sphinx_design==0.6.1",
]

[tool.flit.module]
name = "ansys.scade.git"

[project.urls]
Source = "https://github.com/ansys/scade-git"
Issues = "https://github.com/ansys/scade-git/issues"
Discussions = "https://github.com/ansys/scade-git/discussions"
Documentation = "https://git.scade.docs.pyansys.com"
Releases = "https://github.com/ansys/scade-git/releases"
Homepage = "https://www.ansys.com/products/embedded-software/ansys-scade-suite"

[project.entry-points."ansys.scade.registry"]
srg = "ansys.scade.git:srg"

[project.scripts]
etpmerge = "ansys.scade.git.etpmerge.__main__:main"
almgtmerge = "ansys.scade.git.almgtmerge.__main__:main"
register_ansys_scade_git = "ansys.scade.git.register:main"

[tool.ruff]
line-length = 100
select = [
    "E",    # pycodestyle, see https://beta.ruff.rs/docs/rules/#pycodestyle-e-w
    "D",    # pydocstyle, see https://beta.ruff.rs/docs/rules/#pydocstyle-d
    "F",    # pyflakes, see https://beta.ruff.rs/docs/rules/#pyflakes-f
    "I",    # isort, see https://beta.ruff.rs/docs/rules/#isort-i
    "N",    # pep8-naming, see https://beta.ruff.rs/docs/rules/#pep8-naming-n
]
ignore = [
    "D203",  # One blank line required before class docstring.
    "D204"  # One blank line required after class docstring.
]

[tool.ruff.format]
quote-style = "preserve"

[tool.ruff.isort]
combine-as-imports = true
force-sort-within-sections = true
known-first-party = ["ansys", "conftest", "test_utils"]

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.ruff.lint.per-file-ignores]
"tests/*.py" = ["D",]

[tool.coverage.run]
source = ["ansys.scade.git"]

[tool.coverage.report]
show_missing = true
exclude_also = [
    "raise NotImplementedError",
    "@(abc\\.)?abstractmethod",
    ]

[tool.pytest.ini_options]
testpaths = "tests"
addopts = "-ra --cov=ansys.scade.git --cov-report html:.cov/html --cov-branch --cov-report xml:.cov/xml --cov-report term -vv"

[tool.towncrier]
package = "ansys.scade.git"
directory = "doc/changelog.d"
filename = "doc/source/changelog.rst"
start_string = ".. towncrier release notes start\n"
template = "doc/changelog.d/changelog_template.jinja"
title_format = "`{version} <https://github.com/ansys/scade-git/releases/tag/v{version}>`_ - {project_date}"
issue_format = "`#{issue} <https://github.com/ansys/scade-git/pull/{issue}>`_"

[[tool.towncrier.type]]
directory = "added"
name = "Added"
showcontent = true

[[tool.towncrier.type]]
directory = "changed"
name = "Changed"
showcontent = true

[[tool.towncrier.type]]
directory = "fixed"
name = "Fixed"
showcontent = true

[[tool.towncrier.type]]
directory = "dependencies"
name = "Dependencies"
showcontent = true

[[tool.towncrier.type]]
directory = "miscellaneous"
name = "Miscellaneous"
showcontent = true
[[tool.towncrier.type]]
directory = "documentation"
name = "Documentation"
showcontent = true

[[tool.towncrier.type]]
directory = "maintenance"
name = "Maintenance"
showcontent = true

[[tool.towncrier.type]]
directory = "test"
name = "Test"
showcontent = true
